requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==10.4.0
sqlparse==0.5.2
typing-extensions==4.12.2
